Extract executors into their own module
